chore: split logic per platform
